pyproject.toml
[project]
name = "trellocli"
version = "0.0.1"
authors = [
	{ name = "Elaine Chan", email = "elaineyunru.chan@gmail.com" }
]
description = "CLI program to modify your Trello boards from your computer's command line"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
	"Programming Language :: Python :: 3",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
]
dependencies = [
	"pytest==7.4.0",
    "pytest-mock==3.11.1",
	"py-trello==0.19.0",
	"python-dotenv==1.0.0",
	"cli-test-helpers==3.4.0",
	"typer==0.9.0",
	"rich==13.4.2",
	"simple-term-menu==1.6.1"
]

[project.urls]
"Homepage" = "https://github.com/elainechan01/trellocli"

[project.scripts]
trellocli = "trellocli.__main__:main"
